Another try at getting Windows compiling

+++ tests/src/lib.rs
@@ -25,7 +25,7 @@ mod test {
     unsafe fn shutdown() {
         // Using println or eprintln here will panic as Rust has shut down
         let msg = "We don't test shutdown, but if you see this message it worked!\n";
-        libc::write(1, std::mem::transmute(msg.as_ptr()), msg.len());
+        libc::write(1, std::mem::transmute(msg.as_ptr()), msg.len() as libc::size_t);
     }
 
     #[test]
